pyproject.toml
[project]
name = "agents-learnings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.59.0",
    "pydantic-ai",
    "asyncpg",
    "python-dotenv",
    "logfire",
    "tabulate>=0.9.0",
    "fastapi",
    "uvicorn",
]
